+++ R/input-file.R
@@ -14,8 +14,8 @@
 #'
 #' @inheritParams textInput
 #' @param multiple Whether the user should be allowed to select and upload
-#'   multiple files at once. **Does not work on older browsers, including
-#'   Internet Explorer 9 and earlier.**
+#'   multiple files at once. Support for selecting multiple files depends on
+#'   the browser's support for the HTML `multiple` attribute on file inputs.
 #' @param accept A character vector of "unique file type specifiers" which gives
 #'   the browser a hint as to the type of file the server expects. Many browsers
 #'   use this prevent the user from selecting an invalid file.