+++ R/input-select.R
@@ -13,11 +13,11 @@
 #' `selectInput("letter", "Letter", c("Choose one" = "", LETTERS))`
 #'
 #' **Performance note:** `selectInput()` and `selectizeInput()` can slow down
-#' significantly when thousands of choices are used; with legacy browsers like
-#' Internet Explorer, the user interface may hang for many seconds. For large
-#' numbers of choices, Shiny offers a "server-side selectize" option that
+#' significantly when thousands of choices are used, and the user interface may
+#' hang for many seconds. For large numbers of choices, Shiny offers a
+#' "server-side selectize" option that
 #' massively improves performance and efficiency; see
-#' [this selectize article](https://shiny.rstudio.com/articles/selectize.html)
+#' [this selectize article](https://shiny.posit.co/articles/selectize.html)
 #' on the Shiny Dev Center for details.
 #'
 #' @inheritParams textInput