+++ inst/www/shared/shiny.js
@@ -859,9 +859,6 @@
     return x2;
   }
   function isVisible(el) {
-    if (typeof el.checkVisibility === "function") {
-      return el.checkVisibility();
-    }
     if (el.offsetWidth !== 0 || el.offsetHeight !== 0) {
       return true;
     }
@@ -1164,11 +1161,6 @@
     // htmlOutputBinding.renderValue() (for dynamically-added input objects).
     // This is called before the input is bound.
     initialize(el) {
-      return;
-      el;
-    }
-    // This is called after unbinding the output.
-    dispose(el) {
       return;
       el;
     }
@@ -3074,9 +3066,11 @@
     }
     renderValue(el, data) {
       el.setAttribute("href", data);
-      el.classList.remove("disabled");
-      el.removeAttribute("aria-disabled");
-      el.removeAttribute("tabindex");
+      if (!el.hasAttribute("data-shiny-disable-auto-enable") && !el.classList.contains("shinyjs-disabled")) {
+        el.classList.remove("disabled");
+        el.removeAttribute("aria-disabled");
+        el.removeAttribute("tabindex");
+      }
     }
     // Progress shouldn't be shown on the download button
     // (progress will be shown as a page level pulse instead)
@@ -3090,9 +3084,14 @@
     "click.shinyDownloadLink",
     "a.shiny-download-link",
     function(e4) {
+      const el = e4.currentTarget;
+      if (el.classList.contains("disabled") || !el.getAttribute("href")) {
+        e4.preventDefault();
+        return;
+      }
       const evt = import_jquery25.default.Event("shiny:filedownload");
-      evt.name = this.id;
-      evt.href = this.href;
+      evt.name = el.id;
+      evt.href = el.href;
       (0, import_jquery25.default)(document).trigger(evt);
       return;
       e4;
@@ -6298,6 +6297,7 @@ ${duplicateIdMsg}`;
   var messageHandlers = {};
   var customMessageHandlerOrder = [];
   var customMessageHandlers = {};
+  var conditionalShownClass = "shiny-conditional--shown";
   function addMessageHandler(type, handler) {
     if (messageHandlers[type]) {
       throw 'handler for message of type "' + type + '" already added.';
@@ -6770,15 +6770,15 @@ ${duplicateIdMsg}`;
         const nsPrefix = el.attr("data-ns-prefix");
         const nsScope = this._narrowScope(scope, nsPrefix);
         const show3 = Boolean(condFunc(nsScope));
-        const showing = el.css("display") !== "none";
+        const showing = el.hasClass(conditionalShownClass);
         if (show3 !== showing) {
           if (show3) {
             el.trigger("show");
-            el.show();
+            el.addClass(conditionalShownClass);
             el.trigger("shown");
           } else {
             el.trigger("hide");
-            el.hide();
+            el.removeClass(conditionalShownClass);
             el.trigger("hidden");
           }
         }
@@ -7491,18 +7491,36 @@ ${duplicateIdMsg}`;
       function reportsTheme(el) {
         return el.classList.contains("shiny-image-output") || el.classList.contains("shiny-plot-output") || el.classList.contains("shiny-report-theme");
       }
-      function handleVisualChange(el) {
-        doTriggerResize(el);
-        doSendHiddenState(el);
-        if (reportsSize(el)) doSendSize(el);
-        if (reportsTheme(el)) doSendTheme(el);
+      function refreshOutputInfo(el, initial = false) {
+        if (!initial) doTriggerResize(el);
+        doSendHiddenState(el, initial);
+        if (reportsSize(el)) doSendSize(el, initial);
+        if (reportsTheme(el)) doSendTheme(el, initial);
+      }
+      function refreshThemeOutputs(initial = false) {
+        (0, import_jquery40.default)(".shiny-bound-output").each(function() {
+          const el = this;
+          if (reportsTheme(el)) doSendTheme(el, initial);
+        });
+      }
+      function registerThemeRefreshSignals() {
+        const scheduleThemeInfoRefresh = sendOutputInfoFns.createObserverCallback(
+          100,
+          () => refreshThemeOutputs()
+        );
+        (0, import_jquery40.default)(window).resize(function() {
+          scheduleThemeInfoRefresh();
+        });
+        (0, import_jquery40.default)(document).on("shiny:themechange", function() {
+          scheduleThemeInfoRefresh();
+        });
       }
       function ensureObservers(el) {
         const $el = (0, import_jquery40.default)(el);
         if (!$el.data("shiny-resize-observer")) {
           const onResize = sendOutputInfoFns.createObserverCallback(
             100,
-            () => handleVisualChange(el)
+            () => refreshOutputInfo(el)
           );
           const ro = new ResizeObserver(() => onResize());
           ro.observe(el);
@@ -7512,7 +7530,7 @@ ${duplicateIdMsg}`;
         if (!$el.data("shiny-intersection-observer")) {
           const onIntersect = sendOutputInfoFns.createObserverCallback(
             100,
-            () => handleVisualChange(el)
+            () => refreshOutputInfo(el)
           );
           const io = new IntersectionObserver(() => onIntersect());
           io.observe(el);
@@ -7539,14 +7557,7 @@ ${duplicateIdMsg}`;
           const id = getIdFromEl(el);
           if (id) outputIds.add(id);
           ensureObservers(el);
-          if (!initial) doTriggerResize(el);
-          doSendHiddenState(el, initial);
-          if (reportsSize(el)) {
-            doSendSize(el, initial);
-          }
-          if (reportsTheme(el)) {
-            doSendTheme(el, initial);
-          }
+          refreshOutputInfo(el, initial);
         });
         visibleOutputs.forEach((id) => {
           if (!outputIds.has(id)) {
@@ -7557,6 +7568,7 @@ ${duplicateIdMsg}`;
       }
       doSendOutputInfo(true);
       sendOutputInfoFns.setSendMethod(inputBatchSender, doSendOutputInfo);
+      registerThemeRefreshSignals();
       initialValues[".clientdata_pixelratio"] = pixelRatio();
       (0, import_jquery40.default)(window).resize(function() {
         inputs.setInput(".clientdata_pixelratio", pixelRatio());